Change original WebMediaPlayer signatures from float to double.
This change puts the final signatures containing doubles in place so that
they can be used by Blink code. After this change lands Blink will be updated
to call these methods instead of the temporary xxxFloat() methods.

BUG=227156
TEST=All existing tests still pass. No user visible behavior has changed.

Review URL: https://codereview.chromium.org/13866046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@194624 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/media/android/webmediaplayer_impl_android.h
@@ -66,7 +66,7 @@ class WebMediaPlayerImplAndroid : public WebMediaPlayerAndroid {
   virtual void PlayInternal() OVERRIDE;
   virtual void PauseInternal() OVERRIDE;
   virtual void SeekInternal(base::TimeDelta time) OVERRIDE;
-  virtual float GetCurrentTimeInternal() const OVERRIDE;
+  virtual double GetCurrentTimeInternal() const OVERRIDE;
   virtual void ReleaseResourcesInternal() OVERRIDE;
   virtual void Destroy() OVERRIDE;
   virtual void RequestExternalSurface() OVERRIDE;
@@ -81,7 +81,7 @@ class WebMediaPlayerImplAndroid : public WebMediaPlayerAndroid {
   // The current playing time. Because the mediaplayer is in the browser
   // process, it will regularly update the |current_time_| by calling
   // OnTimeUpdate().
-  float current_time_;
+  double current_time_;
 
   DISALLOW_COPY_AND_ASSIGN(WebMediaPlayerImplAndroid);
 };
